requirements.txt
boto3
opencv-python
Pillow
python-dotenv
pyaudio
anthropic
